Refactor climbing logic and add check to stop climbing
The code changes in this commit refactor the climbing logic in the `CustomMovementComponent` class. The `CheckShouldStopClimbing` function has been added to determine if the character should stop climbing based on the angle of the climbable surface. If the angle is less than or equal to 60 degrees, climbing is stopped. This change improves the accuracy and control of climbing behavior.